+++ packages/amazonq/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.82.0 2025-07-07
+
+- **Bug Fix** Prompt re-authenticate if auto trigger failed with expired token
+
+## 1.81.0 2025-07-02
+
+- **Bug Fix** Stop auto inline completion when deleting code
+
 ## 1.80.0 2025-07-01
 
 - Miscellaneous non-user-facing changes